requirements.txt
autopep8==1.5.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysqlclient==2.0.3
pycodestyle==2.8.0
toml==0.10.2
Werkzeug==2.0.2